move vulnerability digest file from root into remediations
